storage: Made diskType act as a real property.
Drive should be constructed with some specified disk type.
for compatibility reasons constructor parameter is optional
and diskType getter will try to guess actual disk type.

diskType should be authoritative source for blockDev and networkDev.
Earlier blockDev and networkDev properties was a correct way to
determine type of Drive. Now we use diskType for that and
blockDev and networkDev are just convenient method to check
it's value.

Change-Id: Ia4277334c36920bdc275343e3520d2f39851695f
Signed-off-by: Denis Chaplygin <dchaplyg@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1022961